error codes tests moved to separate file. fixtures for current test files moved to separate file, since they are universal for all current tests.
